Allow user to override default mount_point.
I ran into an issue where it was impossible to create a link (either
symbolic or hard) from the fuse mount to a specific area of an
application.

This change addresses this issue by giving the user the ability to
specify a specific path for a mount_point per bucket.

This change does not affect the default functionality.